Updated run.sh

+++ run.sh
@@ -1,10 +1,12 @@
 #!/bin/bash
 
-if [ -f Quera_Test.zip ]; then
+if [[ -f Quera_Test.zip ]]
+then
     rm -f Quera_Test.zip
 fi
 
-if [ -d Quera_Test ]; then
+if [[ -d Quera_Test ]]
+then
     rm -rf Quera_Test
 fi
 
@@ -13,19 +15,27 @@ mkdir Quera_Test
 mkdir Quera_Test/in
 mkdir Quera_Test/out
 
-if [[ -f checker/checker.cpp ]]; then
-    echo "Checker Found!"
+if [[ -f checker/checker.cpp ]]
+then
+    echo "Checker found"
     cp checker/checker.cpp Quera_Test/tester.cpp
 else
-    echo "No Checker"
+    echo "No checker"
 fi
 
 cnt=0
 
 TEST=tests/*.in
 cnt=0
-cp tests/mapping Quera_Test/mapping
-cp subtasks.json Quera_Test/subtasks.json
+if [[ -f tests/mapping && -f subtasks.json ]]
+then
+    echo "Found subtasks"
+    cp tests/mapping Quera_Test/mapping
+    cp subtasks.json Quera_Test/subtasks.json
+else
+    echo "No subtasks"
+fi
+
 for i in `ls $TEST`
 do
     cnt=$(($cnt+1))
@@ -34,23 +44,32 @@ do
     echo "Copying test $cnt"
     cp $i Quera_Test/in/input$cnt.txt
     cp $o Quera_Test/out/output$cnt.txt
-    cat Quera_Test/mapping | sed s/$l/$cnt/ > Quera_Test/mapping.new
-    mv Quera_Test/mapping.new Quera_Test/mapping
+    if [[ -f Quera_Test/mapping ]]
+    then
+        cat Quera_Test/mapping | sed s/$l/$cnt/ > Quera_Test/mapping.new
+        mv Quera_Test/mapping.new Quera_Test/mapping
+    fi
 done
-
-g++ -O2 -std=c++17 -o Quera_Test/jc jc.cpp
 
 cd Quera_Test
 
-./jc
+if [[ -f mapping ]]
+then
+    echo 'Building config.json'
+    g++ -O2 -std=c++17 -o jc ../jc.cpp
+    ./jc
+fi
 
 rm -f mapping subtasks.json jc
 
-if [[ -f config.json && -f tester.cpp ]]; then
+if [[ -f config.json && -f tester.cpp ]]
+then
     zip -r Quera_Test.zip in out tester.cpp config.json
-elif [[ -f config.json ]]; then
+elif [[ -f config.json ]]
+then
     zip -r Quera_Test.zip in out config.json
-elif [[ -f tester.cpp ]]; then
+elif [[ -f tester.cpp ]]
+then
     zip -r Quera_Test.zip in out tester.cpp
 else
     zip -r Quera_Test.zip in out
